moved useInternalSwitcher to global hooks, small folder refactor
