+++ code/src/layouts/intro-banner/IntroBanner.constants.ts
@@ -1,5 +1,5 @@
-import heroImage1 from "@/assets/images/home-page/hero-section-img-1.webp";
-import heroImage2 from "@/assets/images/home-page/hero-section-img-2.webp";
-import heroImage3 from "@/assets/images/home-page/hero-section-img-3.webp";
+import heroImage1 from "@/assets/images/home-page/intro-banner/hero-section-img-1.webp";
+import heroImage2 from "@/assets/images/home-page/intro-banner/hero-section-img-2.webp";
+import heroImage3 from "@/assets/images/home-page/intro-banner/hero-section-img-3.webp";
 
 export const introBannerImages = [heroImage1, heroImage2, heroImage3];
